# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CivicAI – Smart Civic Issue Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List all issues
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, resolved]
        - name: issue_type
          in: query
          schema:
            type: string
            enum: [garbage, pothole, water_leakage, street_light, plastic, other]
        - name: department
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create a new issue report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueBody"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "409":
          description: Duplicate issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get a single issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Issue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update issue status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueBody"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"

  # ─── AI DETECTION ─────────────────────────────────────────────────────────
  /issues/detect:
    post:
      operationId: detectIssue
      tags: [issues]
      summary: AI detection of issue from image data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectIssueBody"
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionResult"

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /users/{id}/issues:
    get:
      operationId: getUserIssues
      tags: [users]
      summary: List issues reported by a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of user issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
...
  /authorities/issues:
    get:
      operationId: getAuthorityIssues
      tags: [authorities]
      summary: Get issues assigned to authority's department
      parameters:
        - name: department
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, resolved]
      responses:
        "200":
          description: Assigned issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"

  # ─── ANALYTICS / DASHBOARD ──────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from /users/login, /authorities/login, or /auth/verify-otp
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    Issue:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        authorityId:
          type: integer
          nullable: true
        issueType:
          type: string
          enum: [garbage, pothole, water_leakage, street_light, plastic, other]
        description:
          type: string
        imageUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        reporterLatitude:
          type: number
          nullable: true
        reporterLongitude:
          type: number
          nullable: true
        address:
          type: string
        status:
          type: string
          enum: [pending, in_progress, resolved]
        verificationStatus:
          type: string
          enum: [pending, flagged, rejected, approved]
        isDuplicate:
          type: boolean
        isValid:
          type: boolean
        validationNotes:
          type: string
          nullable: true
        department:
          type: string
        confidenceScore:
          type: number
        resolvedImageUrl:
          type: string
          nullable: true
        imageHash:
          type: string
        imageMd5:
          type: string
          nullable: true
        imagePhash:
          type: string
          nullable: true
        exifTakenAt:
          type: string
          format: date-time
          nullable: true
        exifLatitude:
          type: number
          nullable: true
        exifLongitude:
          type: number
          nullable: true
        reporterName:
          type: string
          nullable: true
        pointsAwarded:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - issueType
        - latitude
        - longitude
        - status
        - verificationStatus
        - isDuplicate
        - isValid
        - department
        - imageHash
        - pointsAwarded
        - createdAt
        - updatedAt
    CreateIssueBody:
      type: object
      properties:
        userId:
          type: integer
        issueType:
          type: string
          enum: [garbage, pothole, water_leakage, street_light, plastic, other]
        description:
          type: string
        imageUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        reporterLatitude:
          type: number
        reporterLongitude:
          type: number
        address:
          type: string
        confidenceScore:
          type: number
        imageHash:
          type: string
      required: [latitude, longitude, address, imageHash]
    UpdateIssueBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, resolved]
        resolvedImageUrl:
          type: string
    DetectIssueBody:
      type: object
      properties:
        imageBase64:
          type: string
        imageUrl:
          type: string
      required: []
    DetectionResult:
      type: object
      properties:
        issueType:
          type: string
          enum: [garbage, pothole, water_leakage, street_light, plastic, other]
        confidence:
          type: number
        department:
          type: string
        description:
          type: string
      required: [issueType, confidence, department, description]
...